Refactored dart lib for new compount object config.  iOS and Android currently launching without errors

+++ lib/models/background_geolocation.dart
@@ -525,7 +525,7 @@ class BackgroundGeolocation {
       int? timeout,
       int? maximumAge,
       bool? persist,
-      int? desiredAccuracy,
+      double? desiredAccuracy,
       Map<String, dynamic>? extras}) async {
     Map<String, dynamic> options = {};
     if (samples != null) options['samples'] = samples;
